makes a cookie util
